improve experiment zarr managemnt
